+++ Plugin.DeviceInfo/AppState.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+
+namespace Plugin.DeviceInfo
+{
+    public enum AppState
+    {
+        Foreground,
+        Background
+    }
+}