+++ Plugin.DeviceInfo/Platforms/macOS/AppImpl.cs
@@ -8,8 +8,7 @@ namespace Plugin.DeviceInfo
     public partial class AppImpl : IApp
     {
         public bool IsBackgrounded { get; } = false;
-        public IObservable<Unit> WhenEnteringBackground() => Observable.Empty<Unit>();
-        public IObservable<Unit> WhenEnteringForeground() => Observable.Empty<Unit>();
+        public IObservable<AppState> WhenStateChanged() => Observable.Empty<AppState>();
         public IObservable<Unit> EnableIdleTimer(bool enabled) => Observable.Empty<Unit>();
         public bool IsIdleTimerEnabled => true;
     }